+++ api/learning-questions.api.ts
@@ -1,0 +1,6 @@
+import { getMockQuestions } from '@/api/learning-questions.mock';
+import { Question } from '@/types/learning-question';
+
+export async function getQuestions(count: number = 10): Promise<Question[]> {
+  return getMockQuestions(count);
+}
