+++ api/learning-questions.mock.ts
@@ -1,4 +1,4 @@
-import { Question } from '@/types/question';
+import { Question } from '@/types/learning-question';
 
 const mockLearningQuestions: Question[] = [
   {
