+++ app/learning-widget-test/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { getMockQuestions } from '@/api/questions.mock';
+import { getMockQuestions } from '@/api/learning-questions.mock';
 import { LearningWidget } from '@/components/widgets/learning-widget/widget';
 
 export default function Page() {
