+++ components/widgets/learning-widget/widget.tsx
@@ -4,7 +4,7 @@ import { useRef, useState } from 'react';
 import { Button } from '@/components/ui/button';
 import type { CarouselApi } from '@/components/ui/carousel';
 import { Carousel, CarouselContent, CarouselItem, CarouselNext, CarouselPrevious } from '@/components/ui/carousel';
-import { Question } from '@/types/question';
+import { Question } from '@/types/learning-question';
 
 import { FlipCard } from './flip-card';
